requirements.txt
djangorestframework
pillow
PyMySQL
python-decouple
pandas
numpy
django-jazzmin
cryptography